added code and README.md
